31th Cygnus<->FSF merge.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@6996 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cp/errfn.c
@@ -112,7 +112,7 @@ cp_thing (errfn, atarg1, format, arglist)
       if (*f == 'l')
 	++f;
 
-      function = cp_printers[*f];
+      function = cp_printers[(int)*f];
 
       if (function)
 	{
